requirement.txt
clipboard==0.0.4
Pillow==7.1.2
psutil==5.7.0
pyperclip==1.8.0
pywin32==227
selenium==3.141.0
urllib3==1.25.9
